C prototipo/Backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
mysql-connector-python
influxdb-client
python-dotenv
sqlalchemy
pydantic-settings
pymysql
cryptography